+++ bot.py
@@ -4,8 +4,8 @@ import nonebot
 from nonebot.adapters.onebot.v11 import Adapter as ONEBOT_V11Adapter
 
 from src.core.services.dm_provider import init_dm
+from src.core.services.draw_api import close_draw_api_service
 from src.core.services.t2i import close_t2i_service
-
 
 nonebot.init()
 driver = nonebot.get_driver()
@@ -23,6 +23,7 @@ async def _on_startup():
 @driver.on_shutdown
 async def _on_shutdown():
     await close_t2i_service()
+    await close_draw_api_service()
 
 
 if __name__ == "__main__":